requirements.txt
pandas==1.3.5
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==4.3.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Jinja2==2.11.2
MarkupSafe==1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1
python-dotenv==0.20.0
pydash==5.1.0